requirements.txt
streamlit-1.41.1
pinecone-client==2.2.2
sentence-transformers==2.2.2
huggingface-hub==0.4.0
requests==2.31.0